chore: rename type file(index.ts => shared_types.ts)
